# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Hotel Group Internal Procurement MVP API
  version: 0.3.0
  description: >
    Formal MVP API contract covering P0 controls and Phase 1 supplier management,
    procurement request and project initiation. This contract keeps the MVP boundary:
    no CA, e-signature, bid-file encryption/decryption, trusted timestamp,
    tamper-proof evidence, opening hall, or real external system integration.
servers:
  - url: http://127.0.0.1:3000
security:
  - MockUserHeader: []
components:
  securitySchemes:
    MockUserHeader:
      type: apiKey
      in: header
      name: x-mock-user-id
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            auditLogId:
              type: string
  responses:
    Ok:
      description: Request succeeded.
    Created:
      description: Resource created.
    PermissionDenied:
      description: Request rejected by backend policy and recorded with auditLogId when applicable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
paths:
...
  /api/contracts:
    get:
      tags: [Contract Performance]
      summary: Contract ledger list scoped by role and project
      description: Suppliers see only their own enterprise contracts. Buyer, group manager and auditor read by project/organization scope. Expert and admin roles are denied business contract content.
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "403":
          $ref: "#/components/responses/PermissionDenied"
  /api/projects/{projectId}/contracts:
    get:
      tags: [Contract Performance]
      summary: Project contract ledgers
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "403":
          $ref: "#/components/responses/PermissionDenied"
    post:
      tags: [Contract Performance]
      summary: Register contract ledger
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "403":
          $ref: "#/components/responses/PermissionDenied"
  /api/contracts/{contractId}/performance-nodes:
    get:
      tags: [Contract Performance]
      summary: Contract performance nodes
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "403":
          $ref: "#/components/responses/PermissionDenied"
    post:
      tags: [Contract Performance]
      summary: Create performance node
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "403":
          $ref: "#/components/responses/PermissionDenied"
  /api/performance-nodes/{nodeId}/status:
    post:
      tags: [Contract Performance]
      summary: Update performance node status
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "403":
          $ref: "#/components/responses/PermissionDenied"
  /api/contracts/{contractId}/acceptance-payments:
    get:
      tags: [Contract Performance]
      summary: Contract acceptance and payment records
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "403":
          $ref: "#/components/responses/PermissionDenied"
    post:
      tags: [Contract Performance]
      summary: Record acceptance or payment metadata
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "403":
          $ref: "#/components/responses/PermissionDenied"
  /api/contracts/{contractId}/supplier-evaluations:
    post:
      tags: [Contract Performance]
      summary: Create supplier evaluation
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "403":
          $ref: "#/components/responses/PermissionDenied"
  /api/suppliers/{supplierId}/evaluations:
    get:
      tags: [Contract Performance]
      summary: Supplier evaluation history scoped by supplier/project reader
      description: Suppliers see only their own evaluation history. Procurement and audit readers use project/organization scope. Expert and admin roles are denied evaluation content.
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "403":
          $ref: "#/components/responses/PermissionDenied"
...